Update craft-query-api and migrate breaking changes
