+++ composer.json
@@ -4,8 +4,8 @@
   "require": {
     "craftcms/ckeditor": "4.6.0",
     "craftcms/cms": "5.6.9.1",
-    "nystudio107/craft-seomatic": "^5.1.11",
-    "samuelreichor/craft-query-api": "1.1.4",
+    "nystudio107/craft-seomatic": "5.1.11",
+    "samuelreichor/craft-query-api": "2.0.0",
     "samuelreichor/craft-quick-edit": "5.0.2",
     "spacecatninja/imager-x": "5.1.1",
     "vandres/craft-matrix-extended": "^v4.2.1",
