+++ config/project/entryTypes/block_relatedNews--f646c08b-ca3c-41dd-9be9-f173a6175f5e.yaml
@@ -54,10 +54,10 @@ fieldLayouts:
         name: Content
         uid: 47c6e5d9-d36c-48c8-a632-4b23497c8912
         userCondition: null
-handle: block_cta
+handle: block_relatedNews
 hasTitleField: true
 icon: null
-name: 'Block: Cta'
+name: 'Block: Related News'
 showSlugField: true
 showStatusField: true
 slugTranslationKeyFormat: null
