+++ frontend/templates/pages/home.vue
@@ -1,4 +1,11 @@
 <script setup lang="ts">
+import type { EntryRelation } from '~/types/craftcms'
+import type { NewsEntry } from '~/components/RelatedNews/news-type'
+
+export type BlockRelatedNews = {
+  title: string
+  selectNews: EntryRelation[]
+}
 const props = defineProps({
   metadata: {
     type: Object,
@@ -29,10 +36,25 @@ const props = defineProps({
     default: () => undefined,
   },
   relatedNews: {
-    type: Object,
+    type: Object as PropType<BlockRelatedNews>,
     default: () => undefined,
   },
 })
+
+const ids = ref<number[]>([0])
+if (props.relatedNews && props.relatedNews.selectNews.length > 0) {
+  ids.value = props.relatedNews.selectNews.map(news => news.id)
+}
+const currentSite = useCraftCurrentSite()
+const { data: news, error: newsError } = useCraftQuery('entries')
+  .id(ids.value)
+  .siteId(currentSite.value.id)
+  .fields(['headline', 'image', 'richText'])
+  .all()
+
+if (newsError.value) {
+  console.error(newsError.value)
+}
 </script>
 
 <template>
@@ -41,16 +63,18 @@ const props = defineProps({
       v-if="props.image && props.headline"
       :title="props.headline"
       :intro-text="props.richText"
-      :image="props.image[0]"
+      :image="props.image"
       headline-size="h1"
     />
     <CraftArea
       v-if="props.contentBuilder"
       :content="props.contentBuilder"
     />
-    <CraftArea
-      v-if="props.relatedNews"
-      :content="props.relatedNews"
+    <RelatedNews
+      v-if="props.relatedNews && news"
+      :news="news as NewsEntry[]"
+      :headline="props.relatedNews.title"
+      class="mt-20"
     />
   </div>
 </template>
